requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
chroma-hnswlib==0.7.2
chromadb==0.4.7
click==8.1.7
coloredlogs==15.0.1
cryptography==41.0.3
dataclasses-json==0.5.14
exceptiongroup==1.1.3
fastapi==0.99.1
flatbuffers==23.5.26
frozenlist==1.4.0
h11==0.14.0
httptools==0.6.0
humanfriendly==10.0
idna==3.4
importlib-resources==6.0.1
install==1.3.5
langchain==0.0.275
langsmith==0.0.27
marshmallow==3.20.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.5
numpy==1.25.2
onnxruntime==1.15.1
openai==0.27.9
overrides==7.4.0
packaging==23.1
pdfminer.six==20221105
pdfplumber==0.10.2
Pillow==10.0.0
posthog==3.0.2
protobuf==4.24.2
pulsar-client==3.2.0
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.6.3
pypdf==3.15.4
pypdfium2==4.19.0
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
scipy==1.11.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
tiktoken==0.4.0
tokenizers==0.13.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
yarl==1.9.2
